Added container abstraction
I wanted to be able to take advantage of my existing IoC container,
so inspired by NancyFX, I created a container abstraction for NuBot.

To add support for a new container, simply subclass `Container` where
you implement the specific container support, and provide an instance
of that to the `RobotFactory` constructor.

You will still explicitly have to use `RobotFactory.AddPart<T>`, but
I think that's a good thing since it's more of a configuration of the
robot.